# lib/api-zod/src/generated/types/coverLetter.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoverLetter {
  id: string;
  userId: string;
  title: string;
  jobTitle?: string;
  companyName?: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}