# lib/api-zod/src/generated/types/education.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Education {
  degree: string;
  institution: string;
  location?: string;
  graduationYear?: string;
  gpa?: string;
}