# lib/api-zod/src/generated/types/linkedInSummaryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResumeData } from "./resumeData";

export interface LinkedInSummaryBody {
  resumeData: ResumeData;
}