# lib/api-zod/src/generated/types/resumeData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Education } from "./education";
import type { ResumeDataPersonalInfo } from "./resumeDataPersonalInfo";
import type { ResumeDataTemplate } from "./resumeDataTemplate";
import type { WorkExperience } from "./workExperience";

export interface ResumeData {
  personalInfo: ResumeDataPersonalInfo;
  summary?: string;
  workExperience?: WorkExperience[];
  education?: Education[];
  skills?: string[];
  template?: ResumeDataTemplate;
}